% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisR2.r
\name{axisR2}
\alias{axisR2}
\title{Get the \eqn{R^{2}}{R^2} for ordination axes.}
\usage{
axisR2(Data, ord, dist.method = "bray")
}
\arguments{
\item{Data}{A matrix containing only numeric data.}

\item{ord}{The ordination output points dataframe from \code{\link{NMS}}.}

\item{dist.method}{Ecological distance metric to use. Default is Bray-Curtis dissimilarity (\code{"bray"}).}
}
\value{
A dataframe containing the axis \eqn{R^{2}}{R^2} values for the specified ordination.
}
\description{
Compute percent of variation in ordination data explained by each ordination axis.
}
\details{
This function only works for ordinations in 2 or 3 dimensions.

The matrix specified in \code{Data} must contain only numeric values (i.e., no rows or columns of factors, site or species names, etc.) Site and species names may be included as row and column names, however.

Generally \code{ord} will be the \code{"NMSPoints2D.csv"} or \code{"NMSPoints3D.csv"} file created by function \code{\link{NMS}} (see example). However, in theory any matrix of ordination points (generally with samples as rows and ordination axes as columns) should work.

The argument \code{dist.method} passes its value on to \code{\link{vegdist}} in package \code{vegan} and accepts any of its possible values. These include \code{"bray"}, \code{"jaccard"}, \code{"mahalanobis"}, \code{"manhattan"}, \code{"euclidean"}, \code{"binomial"}, and others (see \code{\link{vegdist}} documentation).
}
\examples{
## Create some really simple, random data to look at in ordination space
foo <- matrix(sample(0:20,100,replace=T), ncol = 5, nrow = 20)
colnames(foo) <- paste("Spp", 1:5, sep="")
rownames(foo) <- paste("Site", 1:20, sep="")

## Run an NMS
NMS(foo,only23=TRUE,stepdown=FALSE)

## Read in resultant 2D points file, and get axis R^2 values.
pts2D<-read.csv("NMS Output/NMSPoints2D.csv", row.names = 1, header = TRUE)
axisR2(foo, pts2D)
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}.
}
\seealso{
\code{\link{NMS}}, which runs ordinations on data, and \code{\link{ordRotate}} for rotating ordination axes based on environmental variables. This function wraps and builds upon functionality in \code{\link{vegdist}} from package \code{vegan}.
}
\concept{
relativization, ordination, variation
}

