% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delRare.r
\name{delRare}
\alias{delRare}
\title{Remove rare species from a dataset.}
\usage{
delRare(Data, filt = "perc", level = 0.05)
}
\arguments{
\item{Data}{A matrix containing only numeric data.}

\item{filt}{Specifies the method for determining rare species. Default is percent (\code{"perc"}).}

\item{level}{The level at which species are considered rare. Default is \code{0.05}.}
}
\value{
A dataframe in the same format as the original data. Any rare species (columns) deleted by the function will be removed from this resultant dataset, as will any samples (rows) that are empty after rare species column removal.
}
\description{
Deletes rare species from a dataset based on specified criteria for defining rarity.
}
\details{
Deleting rare species from a dataset is common prior to ordination, as such species often have a large effect on ordination results in spite of generally having little biological effect within the samples or communities in which they are found. A common practice is to remove species occuring in less than 5\% of samples (the defaults for this function). Deleting rare species also commonly has the effect of reducing the \code{\link{cv}} of the dataset. If the dataset is to be also relativized (using, e.g., \code{\link{rel}}), this should be done after removing the rare species, not before. See McCune and Grace (2002) for additional detail.

The matrix specified in \code{Data} must contain only numeric values (i.e., no rows or columns of factors, site or species names, etc.) Site and species names may be included as row and column names, however.

The argument \code{filt} allows rare species determinations to be made according to their presence in a set percentage of samples (\code{"perc"}), in a set number of samples (\code{"num"}), or by their total abundance across all samples (\code{"abund"})

The argument \code{level} should be a proportion (\code{0 < level < 1}) if deleting rare species by the percent of samples in which they appear. Conversely, it should be an integer if deleting rare species by the actual number of samples in which they appear, or by their total abundance across samples.
}
\examples{
## Delete species whose total abundance is < 100 from a simple data matrix
drmat <- delRare(ordmat, filt = "abund", level = 100)
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}.
}
\seealso{
\code{\link{NMS}}, which runs ordinations on data, \code{\link{rel}} for relativizing by rows or columns in the dataset, and \code{\link{cv}} for assessing the coefficients of variaiton of the rows and columns of a dataset.
}
\concept{
relativization, ordination, coefficient of variation, rarity
}

