% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{evar}
\alias{evar}
\title{A simple environmental variable dataframe.}
\format{
A data frame with 20 rows representing sites and 2 columns representing environmental variables.
}
\usage{
evar
}
\description{
A simple environmental variable dataframe.
}
\keyword{datasets}
