% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMS.r
\name{NMS}
\alias{NMS}
\title{Non-metric multidimensional scaling ordination.}
\usage{
NMS(Data, maxruns = 1000, step.maxruns = 100, file.append = "",
  stepdown = TRUE, only23 = FALSE, cores = detectCores() - 1)
}
\arguments{
\item{Data}{A matrix containing only numeric data.}

\item{maxruns}{The maximum number of random starts allowed to find a convergent ordination solution.  Default is \code{1000}.}

\item{step.maxruns}{The number of maxruns used when creating the ordinations for the stepdown plot.  Default is \code{100}.}

\item{file.append}{If specified, appends a character(s) to the end of the code-generated filenames in the output ordination files.}

\item{stepdown}{Should a stepdown series of ordinations, from \code{6} to \code{1} axes, be run and plotted? Default is \code{TRUE}.}

\item{only23}{Should only 2D and 3D ordinations be run (\code{TRUE}), or should R ask if any other dimensions should be tried (\code{FALSE})?  Default is \code{FALSE}.}

\item{cores}{Number of computer cores to parse the function onto. Defaults to \code{1} less than the computer's max.}
}
\value{
Depending on inputs, creates and saves a screeplot showing ordination stress by number of ordination dimensions. It also creates and saves dataframes containing species (columns) and points/sites (rows) ordination values, in the specified number of ordination dimensions. These are saved as \code{.png} and \code{.csv} files in a folder called \code{NMS Output} that the function creates in the working directory.
}
\description{
Run non-metric multidimensional scaling (NMS) ordinations, in several possible dimensions.
}
\details{
Under default settings, this function is rather interactive with the user, requiring input at multiple points. This is intended to allow the user to decide upon the ideal dimensions of the ordination to run after looking at a scree plot, and so on. If you know what you're doing and want instead to run and save an ordination with only a certain number of dimensions, you could specify \code{stepdown = FALSE} and \code{only23 = FALSE}. Note that the latter case is critical if you intend to run \code{NMS} as part of a batch of code; not setting \code{only23} to \code{FALSE} in that case will yield errors.

The matrix specified in \code{Data} must contain only numeric values (i.e., no rows or columns of factors, site or species names, etc.) Site and species names may be included as row and column names, however.

The argument \code{maxruns} corresponds to \code{trymax} in \code{vegan}'s \code{metaMDS} function.

Any text appended to output filenames with \code{file.append} will appear as the last part of the filename, immediately before the \code{.csv} specifying the filetype. Appending such text is generally only useful if you intend to run and compare several competing ordinations (e.g., ordinations with or without rare species included, and so on).

Theoretically, any number of \code{cores} can be specified. However, going above the number of cores present on your computer is generally nonsensical (although the function will still run in that scenario). Fully utilizing all of the computer's cores for this function also makes the computer unavaiable for other uses during the time \code{NMS} is running; thus, the default setting of using \code{1} less than the maximum nuber of available cores is advised.
}
\examples{
## Run an ordination on some simple data, with stepdown and all other defaults
NMS(ordmat)
## Note the output that will be saved in the new "NMS Output" folder on the working directory.

## Plot the ordination 2D results
pts2D <- read.csv("NMS Output/NMSPoints2D.csv", row.names = 1, header = TRUE)
plot(pts2D)
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\seealso{
\code{\link{metaMDS}} in package \code{vegan}, on which this function depends. Pre-processing of the dataframe in preparation for running \code{NMS} may be accomplished using \code{\link{cv}}, \code{\link{delRare}}, and \code{\link{rel}}. Ueful post-processing tools include \code{\link{axisR2}}, and \code{\link{ordRotate}}, plus \code{\link{envfit}} in package \code{vegan}.
}
\concept{
NMS, MDS, NMDS, stress, ordination
}

