% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ordmat}
\alias{ordmat}
\title{A simple matrix for ordination.}
\format{A data frame with 20 rows representing sites and 5 columns representing species.}
\usage{
ordmat
}
\description{
A simple matrix for ordination.
}
\keyword{datasets}

