% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordRotate.r
\name{ordRotate}
\alias{ordRotate}
\title{Rotate ordination axes.}
\usage{
ordRotate(ord, env.fit, env.vector, x.axis = TRUE, positive = TRUE,
  flip = FALSE)
}
\arguments{
\item{ord}{The ordination output points dataframe from \code{\link{NMS}}.}

\item{env.fit}{An \code{\link{envfit}} result object.}

\item{env.vector}{The environmental vector to rotate the ordination about (must be one of the vector names from \code{env.fit}).}

\item{x.axis}{Specifies whether to rotate the ordination along the \code{x} (\code{TRUE}) or \code{y} (\code{FALSE}) axis. Default is \code{TRUE}.}

\item{positive}{Specifies whether to rotate so that the environmental variable is positively (\code{TRUE}) or inversely (\code{FALSE}) related to the axis.  Default is \code{TRUE}.}

\item{flip}{Allows the ordination to be flipped 180 degrees around the other axis.  Default is \code{FALSE}.}
}
\value{
A dataframe in the same form as \code{ord}, with new values generated from the rotated axes.
}
\description{
Rotate a 2D ordination to line up an environmental variable from \code{vegan}'s \code{envfit} with an axis.
}
\details{
This function allows ordination rotation around axes. It is strictly a tool for visual interpretation of the data, with a common practice being to load the variation for the environmental variable of primary interest entirely onto the primary ordination axis. Using this function will not affect the total \eqn{R^{2}}{R^2} of the ordination; however, it will redistribute the variation across the 2 axes. It will also affect the \eqn{R^{2}}{R^2} values of environmental variables overlaid on each axis. Thus, good practice after running this function would be to re-run \code{\link{axisR2}} and \code{\link{envfit}} on these new, rotated points data to generate new \eqn{R^{2}}{R^2} values for axes and environmental variables on the ordination.

This function only works for ordinations in 2 dimensions.
}
\examples{
## Create some really simple, random data to look at in ordination space
foo <- matrix(sample(0:20,100,replace=T), ncol = 5, nrow = 20)
colnames(foo) <- paste("Spp", 1:5, sep="")
rownames(foo) <- paste("Site", 1:20, sep="")

## Run an NMS and read in resultant 2D points file
NMS(foo,only23=TRUE,stepdown=FALSE)
pts2D<-read.csv("NMS Output/NMSPoints2D.csv", row.names = 1, header = TRUE)

## Create some silly environmental variables for envfit
evar <- cbind(sample(0:20, 20, replace = TRUE), sample(3:8, 20, replace = TRUE))
colnames(evar) <- c("Temperature", "pH")
rownames(evar) <- rownames(foo)

## Run envfit
efits <- envfit(pts2D, evar)

## Rotate axis so the temperature variable loads entirely on Axis 1
ordRotate(pts2D, efits, 'Temperature')
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}.
}
\seealso{
\code{\link{NMS}}, which runs ordinations on data, and \code{\link{axisR2}} for the \eqn{R^{2}}{R^2} values of the axes. This function requires and builds upon functionality in \code{\link{envfit}} from package \code{vegan}.
}
\concept{
relativization, ordination, variation, rotation
}

