% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pts2D}
\alias{pts2D}
\title{2-dimensional ordination output.}
\format{
A dataframe with 20 rows representing sites and 2 columns representing ordination axes.
}
\usage{
pts2D
}
\description{
2-dimensional ordination output.
}
\keyword{datasets}
