% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.r
\name{rel}
\alias{rel}
\title{Relativize a dataset.}
\usage{
rel(Data, rel.by = "col")
}
\arguments{
\item{Data}{A matrix containing only numeric data.}

\item{rel.by}{Specifies whether to relativize by column, by row, or both. Defaults to \code{"col"}.}
}
\value{
A dataframe containing relativized values for further ordination analysis.
}
\description{
Relativize a dataset by columns (species) or rows (samples).
}
\details{
Data relativization is generally done prior to running an ordination (using function \code{\link{NMS}}, for example). Relativizing by species (columns), is particularly common, with relativization by samples (rows), somewhat less so. In general relativization should reduce the row and column coefficients of variation (\code{\link{cv}}) of the dataset substantially in order to be worth doing. See McCune and Grace (2002) for additional detail.

The matrix specified in \code{Data} must contain only numeric values (i.e., no rows or columns of factors, site or species names, etc.) Site and species names may be included as row and column names, however.

Relativization can be done by rows (\code{rel.by = "row"}, columns (\code{rel.by = "col"}), or both (\code{rel.by = c("row", "col")} or \code{rel.by = 'both'}).
}
\examples{
## Create some really simple, random data to look at in ordination space
foo <- matrix(sample(0:20,100,replace=T), ncol = 5, nrow = 20)
colnames(foo) <- paste("Spp", 1:5, sep="")
rownames(foo) <- paste("Site", 1:20, sep="")

## Relativize by species maximum
foo2 <- rel(foo, rel.by = "col")
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}. Based on content from Dean Urban and Sarah Goslee.
}
\seealso{
\code{\link{NMS}}, which runs ordinations on data, and \code{\link{cv}} for checking the cv of the rows and columns in the dataset.
}
\concept{
relativization, ordination, coefficient of variation
}

