% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordRotate.r
\name{ordRotate}
\alias{ordRotate}
\title{Rotate ordination axes.}
\usage{
ordRotate(ord, env.fit, env.vector, x.axis = TRUE, positive = TRUE,
  flip = FALSE)
}
\arguments{
\item{ord}{The ordination output points dataframe from \code{\link{NMS}}.}

\item{env.fit}{An \code{\link{envfit}} result object.}

\item{env.vector}{The environmental vector to rotate the ordination about (must be one of the vector names from \code{env.fit}).}

\item{x.axis}{Specifies whether to rotate the ordination along the \code{x} (\code{TRUE}) or \code{y} (\code{FALSE}) axis. Default is \code{TRUE}.}

\item{positive}{Specifies whether to rotate so that the environmental variable is positively (\code{TRUE}) or inversely (\code{FALSE}) related to the axis.  Default is \code{TRUE}.}

\item{flip}{Allows the ordination to be flipped 180 degrees around the other axis.  Default is \code{FALSE}.}
}
\value{
A dataframe in the same form as \code{ord}, with new values generated from the rotated axes.
}
\description{
Rotate a 2D ordination to line up an environmental variable from \code{vegan}'s \code{envfit} with an axis.
}
\details{
This function allows ordination rotation around axes. It is strictly a tool for visual interpretation of the data, with a common practice being to load the variation for the environmental variable of primary interest entirely onto the primary ordination axis. Using this function will not affect the total \eqn{R^{2}}{R^2} of the ordination; however, it will redistribute the variation across the 2 axes. It will also affect the \eqn{R^{2}}{R^2} values of environmental variables overlaid on each axis. Thus, good practice after running this function would be to re-run \code{\link{axisR2}} and \code{\link{envfit}} on these new, rotated points data to generate new \eqn{R^{2}}{R^2} values for axes and environmental variables on the ordination.

This function only works for ordinations in 2 dimensions.
}
\examples{
## Run envfit on some simple environmental data, using the 2-dimensional, NMSPoints2D file (pts2D) output by the NMS function
efits <- envfit(pts2D, evar)

## Rotate axis so the temperature variable loads entirely on Axis 1
rot <- ordRotate(pts2D, efits, 'Temperature')
## Compare rotated ordination to original
par(mfrow = c(2, 1))
plot(pts2D, main = "Original")
plot(rot, main = "Rotated")
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}.
}
\seealso{
\code{\link{NMS}}, which runs ordinations on data, and \code{\link{axisR2}} for the \eqn{R^{2}}{R^2} values of the axes. This function requires and builds upon functionality in \code{\link{envfit}} from package \code{vegan}.
}
\concept{
relativization, ordination, variation, rotation
}

